app/requirements.txt
-e git+git@github.com:tgrbrooks/ADA.git@f97d12c95d34188ecd5781c99ee3fd8c9b248de4#egg=ADA
altgraph==0.17
cycler==0.10.0
ezodf==0.3.2
fbs==0.8.6
future==0.18.2
kiwisolver==1.2.0
lxml==4.6.2
macholib==1.14
matplotlib==3.2.1
numpy==1.18.4
pefile==2019.4.18
pip==9.0.1
PyInstaller==3.4
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.9.2
python-dateutil==2.8.1
pywin32==300
pywin32-ctypes==0.2.0
scipy==1.4.1
setuptools==44.1.1
sip==4.19.8
six==1.15.0
